perf(Container): Do not apply 'styleProps' to parent of children
BREAKING CHANGE: There now is an additional Box
between the Container component and its children.
If you are using the 'styleProps' on the Container to
customize the layout of its children, it is advised to
pass the children inside of a single Box, Flex or Grid
component and to use that component to customize
the layout of the children.